# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SiteContentTrustBadgesItem = {
  label?: string;
};

export interface HeroContent {
  headline: string;
  subheadline: string;
  ctaEstimate: string;
  ctaEmergency: string;
  tag?: string;
}

export interface Service {
  title: string;
  description: string;
  icon: string;
  highlight?: boolean;
}

export interface Review {
  name: string;
  location: string;
  text: string;
  rating: number;
}

export interface MaintenancePlan {
  name: string;
  price: string;
  period: string;
  description?: string;
  features: string[];
  highlight?: boolean;
}

export interface FaqItem {
  question: string;
  answer: string;
}

export interface ContactInfo {
  phone1: string;
  phone2: string;
  address: string;
  website: string;
  license: string;
}

export interface GalleryItem {
  src: string;
  alt: string;
  label: string;
}

export interface SiteContent {
  hero: HeroContent;
  services: Service[];
  reviews: Review[];
  maintenancePlans: MaintenancePlan[];
  trustBadges: SiteContentTrustBadgesItem[];
  faq: FaqItem[];
  serviceAreas: string[];
  contact: ContactInfo;
  gallery: GalleryItem[];
}

export type ContentUpdateValue = { [key: string]: unknown };

export interface ContentUpdate {
  key: string;
  value: ContentUpdateValue;
}

export interface LoginInput {
  password: string;
}

export type Login200 = {
  ok?: boolean;
};

export type GetMe200 = {
  authenticated?: boolean;
};