# lib/api-zod/src/generated/types/contactInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInfo {
  phone1: string;
  phone2: string;
  address: string;
  website: string;
  license: string;
}