# lib/api-zod/src/generated/types/contentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentUpdateValue } from './contentUpdateValue';

export interface ContentUpdate {
  key: string;
  value: ContentUpdateValue;
}