# lib/api-zod/src/generated/types/heroContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HeroContent {
  headline: string;
  subheadline: string;
  ctaEstimate: string;
  ctaEmergency: string;
  tag?: string;
}