# lib/api-zod/src/generated/types/maintenancePlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenancePlan {
  name: string;
  price: string;
  period: string;
  description?: string;
  features: string[];
  highlight?: boolean;
}