# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInfo } from './contactInfo';
import type { FaqItem } from './faqItem';
import type { GalleryItem } from './galleryItem';
import type { HeroContent } from './heroContent';
import type { MaintenancePlan } from './maintenancePlan';
import type { Review } from './review';
import type { Service } from './service';
import type { SiteContentTrustBadgesItem } from './siteContentTrustBadgesItem';

export interface SiteContent {
  hero: HeroContent;
  services: Service[];
  reviews: Review[];
  maintenancePlans: MaintenancePlan[];
  trustBadges: SiteContentTrustBadgesItem[];
  faq: FaqItem[];
  serviceAreas: string[];
  contact: ContactInfo;
  gallery: GalleryItem[];
}